+++ build.gradle.kts
@@ -27,10 +27,10 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-web")
     testImplementation("org.springframework.security:spring-security-test")
     implementation("com.squareup.okhttp3:okhttp:4.11.0")
-    implementation("org.aspectj:aspectjweaver:1.9.20.1")
     implementation("com.auth0:java-jwt:4.4.0")
     implementation(project(":collection"))
     implementation(project(":model"))
+    implementation(project(":source"))
 }
