+++ logger/src/main/java/executor/service/logger/aspect/LoggingAspect.java
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Component;
 public class LoggingAspect {
     private final Logger logger;
 
-    @Before("@within(executor.service.publisher.annotation.Logged)")
+    @Before("@within(executor.service.logger.annotation.Logged)")
     public void logBefore(JoinPoint joinPoint) {
         String methodName = joinPoint.getSignature().getName();
         String className = joinPoint.getTarget().getClass().getSimpleName();
@@ -22,7 +22,7 @@ public class LoggingAspect {
         logger.info(LogMessage.EXECUTING_METHOD.getMessage(), methodName, className, args);
     }
 
-    @AfterReturning(pointcut = "@within(executor.service.publisher.annotation.Logged)", returning = "result")
+    @AfterReturning(pointcut = "@within(executor.service.logger.annotation.Logged)", returning = "result")
     public void logAfterReturning(JoinPoint joinPoint, Object result) {
         String className = joinPoint.getTarget().getClass().getSimpleName();
         String methodName = joinPoint.getSignature().getName();
