+++ settings.gradle.kts
@@ -2,3 +2,4 @@ rootProject.name = "publisher_microservice"
 include("logger")
 include("model")
 include("collection")
+include("source")
