+++ source/src/main/java/executor/service/source/service/proxy/ProxySourceServiceUrl.java
@@ -1,7 +1,7 @@
 package executor.service.source.service.proxy;
 
 import executor.service.model.ProxyConfigHolder;
-import executor.service.model.ProxySource;
+import executor.service.source.model.ProxySource;
 import executor.service.source.okhttp.OkhttpLoader;
 import lombok.RequiredArgsConstructor;
 import okhttp3.Request;
