+++ source/src/main/resources/connection.properties
@@ -1,0 +1,3 @@
+remote.proxy.source=https://freeapiproxies.azurewebsites.net/proxyapi
+remote.proxy.storage=url
+remote.proxy.type=http
