+++ src/main/java/executor/service/processing/proxy/ProxyProcessingServiceImpl.java
@@ -2,7 +2,7 @@ package executor.service.processing.proxy;
 
 import executor.service.exception.validator.UnknownProxyTypeException;
 import executor.service.model.ProxyConfigHolder;
-import executor.service.model.ProxySource;
+import executor.service.source.model.ProxySource;
 import executor.service.collection.queue.proxy.ProxySourceQueueHandler;
 import executor.service.validation.ProxyValidator;
 import org.springframework.stereotype.Component;
