+++ src/main/java/executor/service/processing/proxy/ProxyRemoteProcessingService.java
@@ -1,7 +1,7 @@
 package executor.service.processing.proxy;
 
 import executor.service.processing.RemoteProcessingService;
-import executor.service.model.ProxySource;
+import executor.service.source.model.ProxySource;
 
 public interface ProxyRemoteProcessingService extends RemoteProcessingService<ProxySource> {
 }
