+++ src/main/java/executor/service/processing/proxy/ProxyRemoteProcessingServiceImpl.java
@@ -1,10 +1,10 @@
 package executor.service.processing.proxy;
 
 import executor.service.collection.queue.proxy.ProxySourceQueueHandler;
-import executor.service.exception.source.UnknownSourceServiceException;
 import executor.service.exception.validator.UnknownProxyTypeException;
 import executor.service.model.ProxyConfigHolder;
-import executor.service.model.ProxySource;
+import executor.service.source.exception.UnknownSourceServiceException;
+import executor.service.source.model.ProxySource;
 import executor.service.source.service.proxy.ProxySourceService;
 import executor.service.validation.ProxyValidator;
 import org.springframework.stereotype.Component;
