+++ src/main/resources/application.properties
@@ -1,4 +1,1 @@
 jwt.secret.key=secret
-remote.proxy.source=https://freeapiproxies.azurewebsites.net/proxyapi
-remote.proxy.storage=url
-remote.proxy.type=http
