+++ src/test/java/executor/service/processing/proxy/ProxyProcessingServiceImplTest.java
@@ -1,7 +1,7 @@
 package executor.service.processing.proxy;
 
 import executor.service.exception.validator.UnknownProxyTypeException;
-import executor.service.model.ProxySource;
+import executor.service.source.model.ProxySource;
 import executor.service.model.ProxyConfigHolder;
 import executor.service.collection.queue.proxy.ProxySourceQueueHandler;
 import executor.service.validation.ProxyValidator;
