+++ src/test/java/executor/service/source/service/proxy/ProxySourceServiceFileTest.java
@@ -1,7 +1,7 @@
 package executor.service.source.service.proxy;
 
 import executor.service.model.ProxyConfigHolder;
-import executor.service.model.ProxySource;
+import executor.service.source.model.ProxySource;
 import executor.service.source.reader.FileReader;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
